Fix accursed bug.
The crux was that I was passing a Sphere into `ray_intersect_sphere` *by value*.

This would be fine if the Intersection object returned by the function was purely a position vector. However, the Intersection also contains a reference to the Sphere that was passed into the function. Since it was passed by value, that sphere lived on the stack, and went out of scope when the function returned! This meant any old crap could be written to that memory address at any time!